pyproject.toml
[project]
name = "a2a_assignment"
version = "0.1.0"
dependencies = [
    "a2a-sdk>=0.2.8",
    "fastmcp>=2.9.0",
    "google-generativeai>=0.8.5",
    "pydantic-settings>=2.10.0",
    "streamlit>=1.46.0",
    "uvicorn>=0.34.3",
]
